requirements.txt
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.1.0
langchain-groq>=0.1.0
groq>=0.4.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.10
requests>=2.31.0
streamlit>=1.28.0
openai>=1.0.0
